requirements.txt
ipykernel==6.29.4
jupyterlab==4.2.2
numpy==2.0.0
ollama==0.2.1
openai==1.7.0
pydantic==2.6.3
python-environ==0.4.54

